moved commands to their own modules for clarity
